plane shape for rigid actors
