+++ kool-editor-model/src/commonMain/kotlin/de/fabmax/kool/editor/components/RigidActorComponent.kt
@@ -7,10 +7,7 @@ import de.fabmax.kool.editor.data.RigidActorComponentData
 import de.fabmax.kool.editor.data.RigidActorType
 import de.fabmax.kool.editor.data.ShapeData
 import de.fabmax.kool.editor.model.SceneNodeModel
-import de.fabmax.kool.math.QuatD
-import de.fabmax.kool.math.Vec3d
-import de.fabmax.kool.math.toQuatF
-import de.fabmax.kool.math.toVec3f
+import de.fabmax.kool.math.*
 import de.fabmax.kool.modules.ui2.mutableStateOf
 import de.fabmax.kool.physics.RigidActor
 import de.fabmax.kool.physics.RigidDynamic
@@ -110,14 +107,15 @@ class RigidActorComponent(
         requiredAssets.clear()
         rigidActor?.apply {
             bodyShapes = componentData.properties.shapes
-            geometry = if (bodyShapes.isEmpty()) {
+            val shapes = if (bodyShapes.isEmpty()) {
                 nodeModel.getComponent<MeshComponent>()?.componentData?.shapes
                     ?.mapNotNull { shape -> shape.makeCollisionGeometry() }
                     ?: emptyList()
             } else {
                 bodyShapes.mapNotNull { shape -> shape.makeCollisionGeometry() }
             }
-            geometry.forEach { attachShape(Shape(it)) }
+            shapes.forEach { (shape, pose) -> attachShape(Shape(shape, localPose = pose)) }
+            geometry = shapes.map { it.first }
 
             if (this is RigidDynamic) {
                 updateInertiaFromShapesAndMass()
@@ -128,15 +126,16 @@ class RigidActorComponent(
         setPhysicsTransformFromDrawNode()
     }
 
-    private suspend fun ShapeData.makeCollisionGeometry(): CollisionGeometry? {
+    private suspend fun ShapeData.makeCollisionGeometry(): Pair<CollisionGeometry, Mat4f>? {
         return when (this) {
-            is ShapeData.Box -> BoxGeometry(size.toVec3f())
-            is ShapeData.Capsule -> CapsuleGeometry(length.toFloat(), radius.toFloat())
-            is ShapeData.Cylinder -> CylinderGeometry(length.toFloat(), bottomRadius.toFloat())
-            is ShapeData.Sphere -> SphereGeometry(radius.toFloat())
-            is ShapeData.Heightmap -> loadHeightmapGeometry(this)
+            is ShapeData.Box -> BoxGeometry(size.toVec3f()) to Mat4f.IDENTITY
+            is ShapeData.Capsule -> CapsuleGeometry(length.toFloat(), radius.toFloat()) to Mat4f.IDENTITY
+            is ShapeData.Cylinder -> CylinderGeometry(length.toFloat(), bottomRadius.toFloat()) to Mat4f.IDENTITY
+            is ShapeData.Sphere -> SphereGeometry(radius.toFloat()) to Mat4f.IDENTITY
+            is ShapeData.Heightmap -> loadHeightmapGeometry(this)?.let { it to Mat4f.IDENTITY }
+            is ShapeData.Plane -> PlaneGeometry() to Mat4f.rotation(90f.deg, Vec3f.Z_AXIS)
             is ShapeData.Rect -> null
-            is ShapeData.Empty -> null
+            is ShapeData.Custom -> null
         }
     }
